requirements.txt
torch
torchvision
numpy
onnxruntime
requests
python-dotenv
tqdm
pickle